% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_anonymization.R
\name{adap_buffer_radius_rst}
\alias{adap_buffer_radius_rst}
\title{Title}
\usage{
adap_buffer_radius_rst(data, raster, k, min_distance = 5000)
}
\arguments{
\item{min_distance}{}
}
\description{
Title
}
