% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{addSpace}
\alias{addSpace}
\title{Cpmplete with space}
\usage{
addSpace(char, n)
}
\arguments{
\item{char}{character string}

\item{n}{number of character before sring}
}
\description{
Cpmplete with space
}
