% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{agrisMsg}
\alias{agrisMsg}
\title{generate a message}
\usage{
agrisMsg(txt1, txt2)
}
\arguments{
\item{txt1}{first text}

\item{txt2}{second text}
}
\value{
a character
}
\description{
generate a message
}
