% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\docType{class}
\name{agrisvy-class}
\alias{agrisvy-class}
\title{Create an S4 object that store information about the survey}
\description{
Create an S4 object that store information about the survey
}
\section{Slots}{

\describe{
\item{\code{svyName}}{characterOrNULL.}

\item{\code{author}}{characterOrNULL.}

\item{\code{language}}{characterOrNULL.}

\item{\code{dataDir}}{characterOrNULL.}

\item{\code{varClassDir}}{characterOrNULL.}

\item{\code{preProcScriptDir}}{characterOrNULL.}

\item{\code{preprocDataDir}}{characterOrNULL.}

\item{\code{anoScriptDir}}{characterOrNULL.}

\item{\code{anoDataDir}}{characterOrNULL.}

\item{\code{anoreportDir}}{characterOrNULL.}

\item{\code{fileDesDir}}{characterOrNULL.}

\item{\code{infoLossReport}}{characterOrNULL.}

\item{\code{tempfileDir}}{characterOrNULL.}

\item{\code{aobDir}}{characterOrNULL.}

\item{\code{workingDir}}{characterOrNULL}

\item{\code{enc_args}}{listOrNULL}

\item{\code{rpt_format}}{characterOrNULL}

\item{\code{script_format}}{characterOrNULL}

\item{\code{type}}{characterOrNULL
Class \code{"sdcMicroObj"}

Class to save all information about the survey}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("agrisvy", ...)}.
}

\author{
AMsata Niang
}
\keyword{classes}
