% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_object.R
\docType{class}
\name{agrisvyQuestion-class}
\alias{agrisvyQuestion-class}
\title{An object to model a survey question from Agris survey questionnaire}
\description{
An object to model a survey question from Agris survey questionnaire
}
\section{Slots}{

\describe{
\item{\code{QuestVar}}{name of the variable}

\item{\code{QuestType}}{type of the variable (categorical, continous or text)}

\item{\code{QuestLab}}{text of the question}

\item{\code{QuestValue}}{if categorical, take the different values of the question}

\item{\code{QuestValProbs}}{if categorical, take the probability of each values}
}}

