% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriswksp_object.R
\docType{class}
\name{agriswksp-class}
\alias{agriswksp-class}
\title{Create an S4 object that store information about the survey}
\description{
Create an S4 object that store information about the survey
}
\section{Slots}{

\describe{
\item{\code{wkspName}}{characterOrNULL.}

\item{\code{organizer}}{characterOrNULL.}

\item{\code{riskAnalysisDir}}{characterOrNULL.}

\item{\code{SdcObjDir}}{characterOrNULL.}
}}

