% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoDataDir,agrisvy-method}
\alias{anoDataDir,agrisvy-method}
\title{Goive the complete path of the anoDataDir directory}
\usage{
\S4method{anoDataDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Goive the complete path of the anoDataDir directory
}
