% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoDataDir}
\alias{anoDataDir}
\title{Give the path to the folder containing the anonymized data}
\usage{
anoDataDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the folder containing the anonymized data
}
