% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoreportDir,agrisvy-method}
\alias{anoreportDir,agrisvy-method}
\title{Give the complete path of the anoreportDir directory}
\usage{
\S4method{anoreportDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Give the complete path of the anoreportDir directory
}
