% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoreportDir}
\alias{anoreportDir}
\title{Give the path to the folder containg the anonymization report}
\usage{
anoreportDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the folder containg the anonymization report
}
