% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoScriptDir,agrisvy-method}
\alias{anoScriptDir,agrisvy-method}
\title{Give the complete path of the anoScriptDir directory}
\usage{
\S4method{anoScriptDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\value{
character
}
\description{
Give the complete path of the anoScriptDir directory
}
