% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{anoScriptDir}
\alias{anoScriptDir}
\title{Give the path to the folder coning the anonymization scripts}
\usage{
anoScriptDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the folder coning the anonymization scripts
}
