% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{anova_test_comp}
\alias{anova_test_comp}
\title{Compare one way ANOVA test before and after anonymization}
\usage{
anova_test_comp(design, test = "wilcoxon", alpha = 0.05)
}
\arguments{
\item{design}{survey design specified with the function \code{"agrisvyr::surveyDesign"}}

\item{test}{type of test. \code{"test=c("wilcoxon", "vanderWaerden", "median","KruskalWallis")"}}

\item{alpha}{threshold for the rejection of null hypothesis}
}
\description{
Compare one way ANOVA test before and after anonymization
}
