% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{aobDir,agrisvy-method}
\alias{aobDir,agrisvy-method}
\title{Give the complete path of the aobDir directory}
\usage{
\S4method{aobDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Give the complete path of the aobDir directory
}
