% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{aobDir}
\alias{aobDir}
\title{Give the path of the folder any other business (aob)}
\usage{
aobDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path of the folder any other business (aob)
}
