% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{ArchiveAnoData}
\alias{ArchiveAnoData}
\title{Archive the anonymized data inside an agrisvy object}
\usage{
ArchiveAnoData(agrisvy)
}
\arguments{
\item{agrisvy}{an agrisvy object}
}
\description{
Archive the anonymized data inside an agrisvy object
}
