% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{ArchiveCleanData}
\alias{ArchiveCleanData}
\title{Archive the clean data inside an \code{agrisvy} object}
\usage{
ArchiveCleanData(agrisvy)
}
\arguments{
\item{agrisvy}{an \code{agrisvy} object}
}
\description{
Archive the clean data inside an \code{agrisvy} object
}
