% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{ArchiveProcData}
\alias{ArchiveProcData}
\title{Archive the preprocessed data inside an agrisvy object}
\usage{
ArchiveProcData(agrisvy)
}
\arguments{
\item{agrisvy}{}
}
\description{
Archive the preprocessed data inside an agrisvy object
}
