% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{assignNewVarLabels}
\alias{assignNewVarLabels}
\title{Assign new variable labels define in the Excel file of label in the file description folder}
\usage{
assignNewVarLabels(df, df_lab)
}
\arguments{
\item{df_lab}{}
}
\description{
Assign new variable labels define in the Excel file of label in the file description folder
}
