% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{cat_relation}
\alias{cat_relation}
\title{Provide a matrix to assess dependancy loss due to anonymiziation}
\usage{
cat_relation(obj_i, obj_f, alpha, vars = NULL, df)
}
\arguments{
\item{obj_i}{initial SDCmicro Object}

\item{obj_f}{Final sdcMicro object}

\item{alpha}{dependancy threshold (for p-value)}

\item{vars}{variable to include}
}
\description{
Provide a matrix to assess dependancy loss due to anonymiziation
}
