% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{check_path_limit}
\alias{check_path_limit}
\title{Label binary variable based on pattern in variable names}
\usage{
check_path_limit(agrisvy)
}
\arguments{
\item{agrisvy}{dataframe}
}
\description{
Label binary variable based on pattern in variable names
}
