% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{compare_ind_cat_mean}
\alias{compare_ind_cat_mean}
\title{Compare proportion before and after anonymization}
\usage{
compare_ind_cat_mean(variable, design)
}
\arguments{
\item{design}{}
}
\description{
Compare proportion before and after anonymization
}
