% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{compare_ind_cat_tot}
\alias{compare_ind_cat_tot}
\title{Compare total of categorical variable values before and after anonymization}
\usage{
compare_ind_cat_tot(variable, design)
}
\arguments{
\item{design}{}
}
\description{
Compare total of categorical variable values before and after anonymization
}
