% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{compare_ind_cats}
\alias{compare_ind_cats}
\title{Compare crosstabulation proportion before and after anonymization}
\usage{
compare_ind_cats(variable1, variable2, design)
}
\arguments{
\item{design}{}
}
\description{
Compare crosstabulation proportion before and after anonymization
}
