% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{compare_ind_cats_tot}
\alias{compare_ind_cats_tot}
\title{Compare crosstabulation total before and after anonymization}
\usage{
compare_ind_cats_tot(variable1, variable2, design)
}
\arguments{
\item{design}{}
}
\description{
Compare crosstabulation total before and after anonymization
}
