% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{confint_check_num_by}
\alias{confint_check_num_by}
\title{Compare dissaggregated indicators for numeric variables before and after anonymization}
\usage{
confint_check_num_by(
  design = NULL,
  indicator = "mean",
  numvar = NULL,
  by = NULL
)
}
\arguments{
\item{by}{}
}
\description{
Compare dissaggregated indicators for numeric variables before and after anonymization
}
