% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{copy_data_folder}
\alias{copy_data_folder}
\title{Copy the structure of a folder}
\usage{
copy_data_folder(from, to)
}
\arguments{
\item{from}{directory to copy}

\item{to}{directory where the folder are copied}
}
\description{
Copy the structure of a folder
}
