% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_ano_folders}
\alias{create_ano_folders}
\title{Create anonymization working folders}
\usage{
create_ano_folders(agrisvy, overwrite = FALSE)
}
\arguments{
\item{agrisvy}{}

\item{overwrite}{should existing folders be overwitten}
}
\value{
logical
}
\description{
Create anonymization working folders
}
