% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_ano_r}
\alias{create_ano_r}
\title{Create anonymization script for a data file}
\usage{
create_ano_r(path_to_data, file, pattern)
}
\arguments{
\item{path_to_data}{path to the data folder}

\item{file}{data file for which to create anonymization script}

\item{pattern}{extension of the data file}
}
\value{

}
\description{
Create anonymization script for a data file
}
