% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_folder}
\alias{create_folder}
\title{Create a folder}
\usage{
create_folder(directory, overwrite = FALSE)
}
\arguments{
\item{directory}{directory to create}

\item{overwrite}{if existing directory should be removed}
}
\value{
logical
}
\description{
Create a folder
}
