% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_wb}
\alias{create_wb}
\title{Create an Excel file for variable classification for a given data folder}
\usage{
create_wb(agrisvy, data, wb_file)
}
\arguments{
\item{data}{a dataframe containing information on the file tree}

\item{wb_file}{name the workbook of a data folder}
}
\description{
Create an Excel file for variable classification for a given data folder
}
