% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_totutials.R
\name{create_wksp_folders}
\alias{create_wksp_folders}
\title{Create folders containing working files of the workshop}
\usage{
create_wksp_folders(agriswksp, overwrite = FALSE)
}
\arguments{
\item{agriswksp}{an agriswksp object}

\item{overwrite}{if existing folders should be overwritten}
}
\description{
Create folders containing working files of the workshop
}
