% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{createAgrisvy}
\alias{createAgrisvy}
\title{Create an \code{agrisvy} object}
\usage{
createAgrisvy(
  svyName = "[Survey name]",
  author = "[Author]",
  language = "en",
  workingDir = NULL,
  dataDir = NULL,
  type = NULL
)
}
\arguments{
\item{svyName}{Name of the survey (will be used in the code comments and to customize the anonymization report)}

\item{author}{Author or institution (ex. \code{"FAO"})}

\item{language}{Language (used to customize the language of folder names
and the anonymization report). The possible values are \code{"en"} for English,
\code{"fr"} for French and \code{"es"} for Spanish. The default value is \code{"en"}.}

\item{workingDir}{path to the working directory (where the anonymization working folder will be generated).
It is recommended for this folder to be empty.}

\item{dataDir}{path to the folder where the data to be anonymized are located}

\item{type}{Format of the microdata. The possible values are \code{".dta"} for STATA, \code{".sav"} for SPSS}
}
\value{
an \code{agrisvy} object
}
\description{
An \code{agrisvy} object are used to store useful information that help generate files, folders
scripts but most importantly customize pre-populated scripts and anonymization report
template. It is used as input by some bulk run function for pre-processing, anonymization
archiving, etc.
}
