% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriswksp_object.R
\name{createAgriwksp}
\alias{createAgriwksp}
\title{Create an AGRIS workshop object}
\usage{
createAgriwksp(
  wkspName = "[Workshop name]",
  organizer = "[Organizer]",
  language = "en",
  subtitle = "",
  workingDir = NULL,
  path = NULL,
  type = NULL
)
}
\arguments{
\item{type}{}
}
\description{
Create an AGRIS workshop object
}
