% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_object.R
\name{createQuestion}
\alias{createQuestion}
\title{Title}
\usage{
createQuestion(
  QuestVar = NULL,
  QuestType = NULL,
  QuestLab = NULL,
  QuestValue = NULL,
  QuestValProbs = NULL
)
}
\arguments{
\item{QuestVar}{name of the variable}

\item{QuestType}{type of the variable (categorical, continous or text)}

\item{QuestLab}{text of the question}

\item{QuestValue}{if categorical, take the different values of the question}

\item{QuestValProbs}{if categorical, take the probability of each values}
}
\description{
Title
}
