% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_object.R
\name{createQuestionnaire}
\alias{createQuestionnaire}
\title{Create a questionnaire object (liste of question objects)}
\usage{
createQuestionnaire()
}
\description{
Create a questionnaire object (liste of question objects)
}
