% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{DataPath,agrisvy-method}
\alias{DataPath,agrisvy-method}
\title{Give the complete path to the data to be anonymized}
\usage{
\S4method{DataPath}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\value{
character
}
\description{
Give the complete path to the data to be anonymized
}
