% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{DataPath}
\alias{DataPath}
\title{Contain the path to the data to be anonymized}
\usage{
DataPath(obj)
}
\arguments{
\item{agrissvy}{object of the anonymization}
}
\description{
Contain the path to the data to be anonymized
}
