% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_anonymization.R
\name{dispalce_cluster_df}
\alias{dispalce_cluster_df}
\title{Title}
\usage{
dispalce_cluster_df(
  st_data,
  sf_boundary,
  boundary_name,
  distance_var = "distance",
  progress = FALSE
)
}
\description{
Title
}
