% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{display_miss}
\alias{display_miss}
\title{Summarise missing value due to anonymization}
\usage{
display_miss(obj_i, obj_f, var, df)
}
\arguments{
\item{obj_i}{initial SDC object}

\item{obj_f}{final SDC object}

\item{var}{categorical quasi-identifier}

\item{df}{}
}
\description{
Summarise missing value due to anonymization
}
