% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{.createExcelInfos}
\alias{.createExcelInfos}
\title{Information on Excel files and workbook to be created}
\usage{
.createExcelInfos(agrisvy, collapse_space = TRUE)
}
\arguments{
\item{agrisvy}{agrisurvey object
@importFrom  plyr rbind.fill}
}
\description{
Information on Excel files and workbook to be created
}
