% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{fileDesDir,agrisvy-method}
\alias{fileDesDir,agrisvy-method}
\title{Give the complete path of the fileDesDir directory}
\usage{
\S4method{fileDesDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Give the complete path of the fileDesDir directory
}
