% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{finMessage}
\alias{finMessage}
\title{Print initial message in finalization of anonymization}
\usage{
finMessage(msg)
}
\arguments{
\item{msg}{message to be printed}
}
\description{
Print initial message in finalization of anonymization
}
