% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agrisdata.R
\name{gen_names}
\alias{gen_names}
\title{Generate first name and last name based on sex}
\usage{
gen_names(sex, first_last)
}
\arguments{
\item{sex}{sex for which to generate name takes "Male" or Female}

\item{first_last}{takes "first", "last" or "both"}
}
\description{
Generate first name and last name based on sex
}
\examples{
\dontrun{
gen_names("Male","both")
}
}
