% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_description.R
\name{genAllFileDes}
\alias{genAllFileDes}
\title{Generate file description for the data}
\usage{
genAllFileDes(agrisvy, id_cols = NULL, encoding = "UTF-8")
}
\arguments{
\item{agrisvy}{an agris survey object}

\item{id_cols}{column to ignore}
}
\description{
Generate file description for the data
}
