% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_description.R
\name{genDataFolderDes}
\alias{genDataFolderDes}
\title{Generate a worbook containing description of all data in the folder}
\usage{
genDataFolderDes(agrisvy, data, wb_name, id_cols)
}
\arguments{
\item{agrisvy}{an agris survey object}

\item{data}{data path}

\item{wb_name}{workbook name}

\item{id_cols}{column to exclude}
}
\description{
Generate a worbook containing description of all data in the folder
}
