% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{genDataList}
\alias{genDataList}
\title{Put all data files in a list}
\usage{
genDataList(path, type)
}
\arguments{
\item{path}{path to data folder}

\item{type}{extension of the data}
}
\description{
Put all data files in a list
}
