% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agrisdata.R
\name{generate_agrisdata}
\alias{generate_agrisdata}
\title{generate agris survey data}
\usage{
generate_agrisdata(
  shpfile,
  level = 3,
  geo_id = c("ID_1", "ID_2", "ID_3"),
  geo_name = c("NAME_1", "NAME_2", "NAME_3"),
  np = 30,
  na_prob = 5e-04,
  language = "en"
)
}
\arguments{
\item{language}{}
}
\description{
generate agris survey data
}
