% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_ano_r}
\alias{generate_ano_r}
\title{Generate all anonymization template scripts}
\usage{
generate_ano_r(agrisvy, md = FALSE)
}
\arguments{
\item{md}{}
}
\description{
Generate all anonymization template scripts
}
