% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_preproc_r}
\alias{generate_preproc_r}
\title{Generate preprocessing scripts}
\usage{
generate_preproc_r(path_to_data, data_format = "stata")
}
\arguments{
\item{path_to_data}{path to data files}

\item{data_format}{data format}
}
\description{
Generate preprocessing scripts
}
