% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_report_template}
\alias{generate_report_template}
\title{Generate the template of the sdc report}
\usage{
generate_report_template(agrisvy, type)
}
\arguments{
\item{type}{}
}
\description{
Generate the template of the sdc report
}
