% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_varclas}
\alias{generate_varclas}
\title{Create Excel file for variable classification for a given data folder}
\usage{
generate_varclas(path_to_data, data_format = "stata")
}
\arguments{
\item{path_to_data}{path to data folder}

\item{data_format}{format of the data (default is stata)}
}
\value{

}
\description{
Create Excel file for variable classification for a given data folder
}
