% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_totutials.R
\name{generateLabFiles}
\alias{generateLabFiles}
\title{Generate needed file sfor wkrshop practice}
\usage{
generateLabFiles(agrisvy, overwrite)
}
\arguments{
\item{overwrite}{}
}
\description{
Generate needed file sfor wkrshop practice
}
