% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_description.R
\name{genFileDes}
\alias{genFileDes}
\title{Generate file description for a given dataset}
\usage{
genFileDes(file, name, id_cols, wb, type)
}
\arguments{
\item{file}{dataset}

\item{name}{name of the sheet where the description should be saved}

\item{id_cols}{column for witch description will not be generated}

\item{wb}{name of the workbook}

\item{type}{format of the data}
}
\description{
Generate file description for a given dataset
}
