% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{GlobRiskTab}
\alias{GlobRiskTab}
\title{Title}
\usage{
GlobRiskTab(sdc, df = FALSE, title = "")
}
\arguments{
\item{sdc}{an object of type SdcMicroObj}

\item{df}{if the function return a dataframe or styled datafram for rmarkdown display}

\item{title}{title of the table in case df=FALSE}
}
\description{
Title
}
