% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{HierRiskSummary}
\alias{HierRiskSummary}
\title{summary of hierarchical risk}
\usage{
HierRiskSummary(sdc, dfl = FALSE, title = "")
}
\arguments{
\item{sdc}{an sdcMicro Object}

\item{dfl}{return a simple dataframe or a styled dataframe}

\item{title}{title of the styled dataframe}
}
\value{
a data frame
}
\description{
summary of hierarchical risk
}
