% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{infoLossReport,agrisvy-method}
\alias{infoLossReport,agrisvy-method}
\title{Give the complete path of the infoLossReport directory}
\usage{
\S4method{infoLossReport}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Give the complete path of the infoLossReport directory
}
