% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{infoLossReport}
\alias{infoLossReport}
\title{Give the path to the folder containing the information loss report}
\usage{
infoLossReport(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the folder containing the information loss report
}
