% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_anonymization.R
\name{k_ano_adaptative_buffer}
\alias{k_ano_adaptative_buffer}
\title{Title}
\usage{
k_ano_adaptative_buffer(
  data,
  raster,
  k,
  min_distance = 5000,
  boundary,
  boundary_name
)
}
\description{
Title
}
