% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_anonymization.R
\name{kAno_adap_buffer0}
\alias{kAno_adap_buffer0}
\title{Title}
\usage{
kAno_adap_buffer0(
  data,
  raster,
  k,
  min_distance = 5000,
  sf_boundary,
  boundary_name,
  progress = TRUE
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
