% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{KanoTab}
\alias{KanoTab}
\title{data frame containing k-anonymity information}
\usage{
KanoTab(sdcObj, df = FALSE, levels = c(2, 3, 5), time, obj)
}
\arguments{
\item{sdcObj}{an sdcMicro object}

\item{df}{logical. retunrn a dataframe or a styled dataframe}

\item{levels}{k-anonymity levels}

\item{title}{title of the table in case df=FALSE}
}
\description{
data frame containing k-anonymity information
}
