% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{label_binary}
\alias{label_binary}
\title{Lebeling binary variable}
\usage{
label_binary(data, var, bin = "01", lang = "en")
}
\arguments{
\item{data}{dataset}

\item{var}{binary variable to be labeled}

\item{bin}{takes \code{"01"} or \code{"12"}}

\item{lang}{takes \code{"en"} for English or \code{"fr"} for french}
}
\description{
Lebeling binary variable
}
