% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{label_binary_at}
\alias{label_binary_at}
\title{Label a set of binary labels}
\usage{
label_binary_at(data, vars, bin = "01", lang)
}
\arguments{
\item{data}{dataframe}

\item{vars}{list of binary labels to be labeled}

\item{bin}{takes \code{"01"} or \code{"12"}}

\item{lang}{takes \code{"en"} or \code{"fr"}}
}
\description{
Label a set of binary labels
}
