% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{label_binary_pattern}
\alias{label_binary_pattern}
\title{Label binary variable based on pattern in variable names}
\usage{
label_binary_pattern(data, pattern, bin = "01", lang)
}
\arguments{
\item{data}{dataframe}

\item{pattern}{common pattern of binary variables}

\item{bin}{takes \code{"01"} or \code{"12"}}

\item{lang}{takes \code{"en"} for english or \code{"fr'} for french}
}
\description{
Label binary variable based on pattern in variable names
}
