% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agrisdata.R
\name{label_geo_name}
\alias{label_geo_name}
\title{Label a geo name}
\usage{
label_geo_name(data, label, level, agris_var)
}
\arguments{
\item{data}{geo ref data}

\item{label}{label}

\item{level}{level}

\item{agris_var}{agris var}
}
\description{
Label a geo name
}
