% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{label_val_gen}
\alias{label_val_gen}
\title{Generate a function to label specific variable}
\usage{
label_val_gen(labels, levels)
}
\arguments{
\item{levels}{}
}
\description{
Generate a function to label specific variable
}
