% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{labels}
\alias{labels}
\title{Extract label of all variable of a dataset}
\usage{
labels(fileName, type, encoding = encoding, password, rounds, size)
}
\arguments{
\item{fileName}{dataset name}

\item{type}{data type}
}
\value{
dataframe
}
\description{
Extract label of all variable of a dataset
}
