% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{lm_compare}
\alias{lm_compare}
\title{Compare linear model results before and after anonymization}
\usage{
lm_compare(formula, raw_data, ano_data, ...)
}
\arguments{
\item{...}{}
}
\description{
Compare linear model results before and after anonymization
}
