% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{loadAgrisvyObj}
\alias{loadAgrisvyObj}
\title{Load the agrisvy object in the R session by executing the  script '_R/_setup.R'}
\usage{
loadAgrisvyObj()
}
\description{
Load the agrisvy object in the R session by executing the  script '_R/_setup.R'
}
