% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_anonymization.R
\name{Mk_ano_adaptative_buffer}
\alias{Mk_ano_adaptative_buffer}
\title{Title}
\usage{
Mk_ano_adaptative_buffer(
  data,
  data_id_var,
  raster,
  k_cluster,
  com_feature = NULL,
  k_com = NULL,
  min_distance = 5000,
  boundary,
  boundary_name,
  crop_buffer = TRUE,
  logical_comb = "&"
)
}
\arguments{
\item{logical_comb}{}
}
\description{
Title
}
