% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{preprocDataDir,agrisvy-method}
\alias{preprocDataDir,agrisvy-method}
\title{give the complete path of the preprocDataDir directory}
\usage{
\S4method{preprocDataDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
give the complete path of the preprocDataDir directory
}
