% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{preprocDataDir}
\alias{preprocDataDir}
\title{Give the path to folder containing the pre-processed data}
\usage{
preprocDataDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to folder containing the pre-processed data
}
