% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{preprocMessage}
\alias{preprocMessage}
\title{Print a initial info message during preprocessing}
\usage{
preprocMessage(msg)
}
\arguments{
\item{msg}{message to be printed}
}
\description{
Print a initial info message during preprocessing
}
