% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{preProcScriptDir,agrisvy-method}
\alias{preProcScriptDir,agrisvy-method}
\title{give the complete path of the preProcScriptDir directory}
\usage{
\S4method{preProcScriptDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\value{
character
}
\description{
give the complete path of the preProcScriptDir directory
}
