% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{preProcScriptDir}
\alias{preProcScriptDir}
\title{Give the path to the folder containing the scripts of pre-processing}
\usage{
preProcScriptDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the folder containing the scripts of pre-processing
}
