% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{RenderGlobalRisk,sdcReportObj-method}
\alias{RenderGlobalRisk,sdcReportObj-method}
\title{render global risk summary in Rmarkdown}
\usage{
\S4method{RenderGlobalRisk}{sdcReportObj}(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\description{
render global risk summary in Rmarkdown
}
