% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{RenderGlobalRisk}
\alias{RenderGlobalRisk}
\title{Render global risk summary in Rmarkdown ducument}
\usage{
RenderGlobalRisk(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\description{
Render global risk summary in Rmarkdown ducument
}
