% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{renderHierRiskSummary,sdcReportObj-method}
\alias{renderHierRiskSummary,sdcReportObj-method}
\title{render hierarchical risk summary}
\usage{
\S4method{renderHierRiskSummary}{sdcReportObj}(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\value{
a dataframe
}
\description{
render hierarchical risk summary
}
