% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{renderHierRiskSummary}
\alias{renderHierRiskSummary}
\title{render hierarchical risk summary}
\usage{
renderHierRiskSummary(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\value{
a dataframe
}
\description{
render hierarchical risk summary
}
