% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{renderKanoTab}
\alias{renderKanoTab}
\title{render k-anonymity table in rmarkdown}
\usage{
renderKanoTab(obj, levels = c(2, 3, 5), time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\value{
a dtaframe
}
\description{
render k-anonymity table in rmarkdown
}
