% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop_totutials.R
\name{renderPreprocFiles}
\alias{renderPreprocFiles}
\title{Render html, pdf or MS word file or the pre-processing files}
\usage{
renderPreprocFiles(agrisvy, type, format)
}
\arguments{
\item{agrisvy}{an \code{agriswkp} object}

\item{type}{}

\item{format}{format of the output (`html`, `pdf`or `docx`)}
}
\description{
Render html, pdf or MS word file or the pre-processing files
}
