% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{renderRiskIndSUmmary,sdcReportObj-method}
\alias{renderRiskIndSUmmary,sdcReportObj-method}
\title{render summary of individual risk}
\usage{
\S4method{renderRiskIndSUmmary}{sdcReportObj}(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\description{
render summary of individual risk
}
