% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{renderRiskIndSUmmary}
\alias{renderRiskIndSUmmary}
\title{Render summary of individual risk}
\usage{
renderRiskIndSUmmary(obj, time = "initial")
}
\arguments{
\item{sdcReportObj}{}
}
\description{
Render summary of individual risk
}
