% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace structural missing value by a given level}
\usage{
replace_missing(data, var, replacement = 999, label = "structural Missing")
}
\arguments{
\item{data}{a dataframe.}

\item{var}{factor column containing structural missing}

\item{replacement}{an integer value of the replacement value}

\item{label}{label of the structural missing value}
}
\value{
a dataframe
}
\description{
Replace structural missing value by a given level
}
