% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriswksp_object.R
\name{riskAnaDir,agriswksp-method}
\alias{riskAnaDir,agriswksp-method}
\title{give the complete path of the preprocDataDir directory}
\usage{
\S4method{riskAnaDir}{agriswksp}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
give the complete path of the preprocDataDir directory
}
