% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{RiskIndSUmmary}
\alias{RiskIndSUmmary}
\title{a  dataframe containing the summary of individual risk}
\usage{
RiskIndSUmmary(sdc, df = FALSE, title = "")
}
\arguments{
\item{sdc}{an sdcMicro Object}

\item{df}{retunrn a simple dataframe or a styled dataframe}

\item{title}{title of the styled dataframe}
}
\value{
a dataframe
}
\description{
a  dataframe containing the summary of individual risk
}
