% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{runPreproc}
\alias{runPreproc}
\title{Run all the pre-processing scripts}
\usage{
runPreproc(agrisvy)
}
\arguments{
\item{agrisvy}{an agris survey objects}
}
\value{
logical
}
\description{
Run all the pre-processing scripts
}
