% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\name{saveReprtObj}
\alias{saveReprtObj}
\title{Title}
\usage{
saveReprtObj(
  agrisvy,
  intialObj = NULL,
  finalObj = NULL,
  unit = NULL,
  hierarchy = NULL,
  global = TRUE,
  individual = TRUE,
  suda = FALSE,
  hierarchical = FALSE,
  childName = NULL
)
}
\arguments{
\item{childName}{}
}
\description{
Title
}
