% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{sdcMessage}
\alias{sdcMessage}
\title{Print initial info message during anonymization}
\usage{
sdcMessage(msg)
}
\arguments{
\item{msg}{message to be printed}
}
\description{
Print initial info message during anonymization
}
