% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_object.R
\docType{class}
\name{sdcReportObj-class}
\alias{sdcReportObj-class}
\title{Create an object containing information for sdc report}
\description{
Create an object containing information for sdc report
}
\section{Slots}{

\describe{
\item{\code{intialObj}}{sdcmicroOrNULL.}

\item{\code{finalObj}}{sdcmicroOrNULL.}

\item{\code{unit}}{characterOrNULL.}

\item{\code{hierarchy}}{characterOrNULL.}

\item{\code{globalRisk}}{logicalOrNULL.}

\item{\code{indRisk}}{logicalOrNULL.}

\item{\code{sudaRisk}}{logicalOrNULL.}

\item{\code{hierRisk}}{logicalOrNULL.}
}}

