% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R, R/agriswksp_object.R
\name{show,agrisvy-method}
\alias{show,agrisvy-method}
\title{Print agrisvy object}
\usage{
\S4method{show}{agrisvy}(object)

\S4method{show}{agrisvy}(object)
}
\arguments{
\item{agrisvy}{an \code{agrisvy} object}
}
\description{
Print agrisvy object

Title
}
