% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{showFolderTree}
\alias{showFolderTree}
\title{Print folder file in tree format}
\usage{
showFolderTree(agrisvy, folders_path)
}
\arguments{
\item{path_folder}{path to the folder}
}
\description{
Print folder file in tree format
}
