% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{surveyDesign}
\alias{surveyDesign}
\title{Specify sample design for information loss asessment}
\usage{
surveyDesign(obj_i, obj_f, ...)
}
\arguments{
\item{obj_i}{initial \code{"sdcMicroObj"} object}

\item{obj_f}{final \code{"sdcMicroObj"} object}

\item{...}{argument of the function \code{"svydesing"} from the package \code{"survey"} (exept the \code{"data"} arguement)}
}
\description{
Specify sample design for information loss asessment
}
