% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{tempfileDir,agrisvy-method}
\alias{tempfileDir,agrisvy-method}
\title{Give the complete path of the tempfileDir directory}
\usage{
\S4method{tempfileDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\description{
Give the complete path of the tempfileDir directory
}
