% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{tempfileDir}
\alias{tempfileDir}
\title{Give the path of the folder containing temporary files generated during the anonymization process}
\usage{
tempfileDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path of the folder containing temporary files generated during the anonymization process
}
