% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{update_labels}
\alias{update_labels}
\title{Extract data label, variable labels, value labels from excel file}
\usage{
update_labels(dat, agrisvy, dataset_name, subfolder = NULL)
}
\description{
Extract data label, variable labels, value labels from excel file
}
