% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{varClassDir,agrisvy-method}
\alias{varClassDir,agrisvy-method}
\title{give the complete path of the varClassDir directory}
\usage{
\S4method{varClassDir}{agrisvy}(obj)
}
\arguments{
\item{agrisvy}{}
}
\value{
character
}
\description{
give the complete path of the varClassDir directory
}
