% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{varClassDir}
\alias{varClassDir}
\title{Give the path to the variable classification folder}
\usage{
varClassDir(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path to the variable classification folder
}
