% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{writeDataFunc}
\alias{writeDataFunc}
\title{Specify haven write function depending on the data type}
\usage{
writeDataFunc(agrisvy)
}
\arguments{
\item{agrisvy}{an agrisvy object}
}
\description{
Specify haven write function depending on the data type
}
