% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{ArchiveAnoData}
\alias{ArchiveAnoData}
\title{Archive the anonymized data inside the AGRIS survey object}
\usage{
ArchiveAnoData(agrisvy)
}
\arguments{
\item{agrisvy}{an \code{agrisvy} object}
}
\description{
Archive the anonymized data inside the AGRIS survey object
}
