% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_ano_r}
\alias{create_ano_r}
\title{Create anonymization script for a data file}
\usage{
create_ano_r(agrisvy, file)
}
\arguments{
\item{file}{data file for which to create anonymization script}
}
\description{
Create anonymization script for a data file
}
