% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{create_preproc_r}
\alias{create_preproc_r}
\title{Create a simple  preprocessing R script for a given dataset}
\usage{
create_preproc_r(path_to_data, file, pattern)
}
\arguments{
\item{path_to_data}{path to the data to be anonymized}

\item{file}{name of the data file}

\item{pattern}{extension of the data file}
}
\description{
Create a simple  preprocessing R script for a given dataset
}
