% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{createAgrisvy}
\alias{createAgrisvy}
\title{Create an agrisvy object}
\usage{
createAgrisvy(
  svyName = "[Survey name]",
  author = "[Author]",
  language = "en",
  workingDir = NULL,
  path = NULL,
  type = NULL
)
}
\arguments{
\item{svyName}{name of the survey}

\item{author}{author (institution)}

\item{language}{language (display in predefined scripts)}

\item{workingDir}{working directory (where the folders will be created)}

\item{path}{path do the microdata of the survey}

\item{type}{extension of the micro data example ".dta" for stata, ".sav" for SPSS}

\item{varClassDir}{name of the variable classification folder}

\item{preProcScriptDir}{name of the pre-processing scripts folder}

\item{preprocDataDir}{name of the preprocessed data folder}

\item{anoScriptDir}{name of the anonymization scripts folder}

\item{anoDataDir}{name of the final anonymized data folder}

\item{anoreportDir}{name of the anonymization report folder}

\item{fileDesDir}{name of the file description folder}

\item{infoLossReport}{name of the information loss report folder}

\item{tempfileDir}{name of the temporary file folder}

\item{aobDir}{name of the miscellenous folder}
}
\value{
a agrisvy object
 @importFrom methods new
}
\description{
Create an agrisvy object
}
