% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{DataPath}
\alias{DataPath}
\title{Give the path of the folder containing the microdata to be anonymized}
\usage{
DataPath(obj)
}
\arguments{
\item{obj}{an \code{agrisvy} object}
}
\value{
a character
}
\description{
Give the path of the folder containing the microdata to be anonymized
}
