% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrisvy_object.R
\name{genDataList}
\alias{genDataList}
\title{Put the different microdata in a list}
\usage{
genDataList(path, type)
}
\arguments{
\item{path}{Path to the folder containing the data}

\item{type}{type of the data. The possible values are \code{".dta"} for STATA and \code{".sav"} for SPSS}
}
\value{
a list
}
\description{
Put the different microdata in a list
}
