% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_agrisdata.R
\name{generate_agrisdata}
\alias{generate_agrisdata}
\title{Generate a sample geo-reference data of agris survey based on the generic questionnaire}
\usage{
generate_agrisdata(n = 60, na_prob = 5e-04)
}
\arguments{
\item{n}{size of sample point per geographic unit}

\item{na_prob}{proportion of random missing data introduced in the data}
}
\description{
Generate a sample geo-reference data of agris survey based on the generic questionnaire
}
