% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_ano_r}
\alias{generate_ano_r}
\title{Generate all anonymization template scripts}
\usage{
generate_ano_r(path_to_data, data_format = "stata")
}
\arguments{
\item{path_to_data}{path to the data folder}

\item{data_format}{data format}
}
\description{
Generate all anonymization template scripts
}
