% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_preproc_r}
\alias{generate_preproc_r}
\title{Generate preprocessing scripts}
\usage{
generate_preproc_r(agrisvy)
}
\arguments{
\item{agrisvy}{}
}
\description{
Generate preprocessing scripts
}
