% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_report_template}
\alias{generate_report_template}
\title{Generate the template of the sdc report}
\usage{
generate_report_template(
  directory = getwd(),
  svy_name = "urvey",
  author = "Amsata"
)
}
\arguments{
\item{directory}{directory of anonymization folders}

\item{svy_name}{name of the survey}

\item{author}{institution}
}
\value{

}
\description{
Generate the template of the sdc report
}
