% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{generate_varclas}
\alias{generate_varclas}
\title{Create Excel file for variable classification for a given data folder}
\usage{
generate_varclas(agrisvy)
}
\arguments{
\item{agrisvy}{}
}
\description{
Create Excel file for variable classification for a given data folder
}
