% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{labels}
\alias{labels}
\title{Extract label of all variable of a dataset}
\usage{
labels(fileName)
}
\arguments{
\item{fileName}{dataset name}
}
\value{

}
\description{
Extract label of all variable of a dataset
}
