% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{miss_values_sum}
\alias{miss_values_sum}
\title{SUmmarise missing value introduced by anonymization}
\usage{
miss_values_sum(obj_i, obj_f)
}
\arguments{
\item{obj_i}{initial sdcMicro object}

\item{obj_f}{final sdcMicro Object}
}
\description{
SUmmarise missing value introduced by anonymization
}
