% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{readDataFunc}
\alias{readDataFunc}
\title{specify haven read function depending on the data type}
\usage{
readDataFunc(agrisvy, path, enquote = TRUE)
}
\arguments{
\item{agrisvy}{an agrisvy object}

\item{enquote}{}
}
\description{
specify haven read function depending on the data type
}
