% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{setup_anonymization}
\alias{setup_anonymization}
\title{set up the anonymization working directory by creating folders and files}
\usage{
setup_anonymization(agrisvy, overwrite)
}
\arguments{
\item{agrisvy}{}

\item{overwrite}{overwrite existing folders}
}
\description{
set up the anonymization working directory by creating folders and files
}
