% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_setup.R
\name{setup_anonymization}
\alias{setup_anonymization}
\title{set up the anonymization working directory by creating folders and files}
\usage{
setup_anonymization(
  agrisvy,
  overwrite = FALSE,
  open = FALSE,
  encoding = "UTF-8"
)
}
\arguments{
\item{agrisvy}{an \code{agrisvy} object}

\item{overwrite}{a \code{logical}. If \code{TRUE},overwrite the working folders
if they already exists.}

\item{encoding}{}
}
\description{
This fucntion set up the anonymization working space by creating the folders,
generating the pre-populated R scripts, generating the anonymization report,
and running the first round of pre-processing and anonymization to make sure
that the process is smooth.
}
\examples{
\dontrun{
agrissvy_obj=createAgrisvy(
                svyName = "AGRIS SURVEY 2023",
                author = "AgriSurvey Team",
                language = "en",
                workingDir = "C/Documents/anonymization",
                dataDir = "C/Documents/AgrisData",
                type = ".dta"
)


}
}
