% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_loss.R
\name{surveyDesign}
\alias{surveyDesign}
\title{Specify sample design for information loss asessment}
\usage{
surveyDesign(x, y, df = FALSE, ...)
}
\arguments{
\item{df}{}
}
\description{
Specify sample design for information loss asessment
}
