% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_func.R
\name{writeDataFunc}
\alias{writeDataFunc}
\title{Specify haven write function depending on the data type}
\usage{
writeDataFunc(agrisvy, dat, path, enquote = TRUE, obj_name = NULL)
}
\arguments{
\item{agrisvy}{an agrisvy object}

\item{obj_name}{}
}
\description{
Specify haven write function depending on the data type
}
